Link to demo document instead of including it
Previously, we included the PDF document generated by the demo example
in the Git repository and referenced it from the readme.  Now we have a
script that builds the example and publishes it to genpdf-rs.ireas.org,
see https://git.sr.ht/~ireas/genpdf-rs-site, so we link to that PDF file
instead.  As the currently published version still points to the
included PDF file, we keep it in the repository until the next release.