Release v0.2.0-alpha.1
This is an alpha version for a minor release that improves the font
handling, adds support for embedding images and contains many small
improvements and bugfixes.

Compared to the previous alpha version, this version contains a fix in
the bounding box calculation for images and a new error message for
images with transparency.