Remove demo.pdf from version control
With commit 1181e23fbc3ed751577b8dbf7208ccdc21e7b72e, we link to a
version of the demo document that has been published to a web page
instead of the demo.pdf in this repository.  As we now have released a
new version, we can remove demo.pdf from version control.