Add page reference to layer and refactor lifetimes
This patch adds a reference to the current Page to Layer.  To make this
possible, we no longer store the Layer instances in Page but the
PdfLayerReference.  We then create Layer instances once they are
required by cloning the PdfLayerReference.  This also means that we can
store the Layer by value in Area instead of using a reference, cloning
it if necessary.

We use this occasion to simplify the lifetimes of TextSection too, which
can clone its area instead of storing a reference to it.