Add methods to access the next layer of a page
This patch adds the Layer::next and Area::next_layer methods that make
it possible to access the next layer of a page.  If this layer already
exists, it is re-used.  Otherwise it is created.

To avoid requiring mutable references to Page, we introduce the new
Layers struct that stores the layer Vec in a RefCell.