Use ascent for vertical text positioning
Previously, we used the glyph height to calculate the vertical position
of a line of text.  This meant that the baseline was too low, leading to
overlaps if text with different sizes was printed next to each other.
With this patch, we fix this issue by using the ascent instead of the
glyph height for the vertical positioning.